Update 1. Importing and Processing Data.R
I added the code here that will exclude the photos that were tagged as "out of range", in most cases when the camera was covered in snow or if the camera changed it's field of view so wasn't recording the same area any more. This will help the model take into account that some of the low winter detections are also because the camera was covered. This will adjust your effort column to only number of days the camera was actually taking photos.